Issue 1152: existingJavaType takes precedence over $ref
In the scenario where both existingJavaType and $ref are set on the same schema node, the creator of the schema intends to rely on an existing Java type for the code generation, but still wants to refer a schema for documentation and interoperability purposes.
Therefore, from code generation standpoint, existingJavaType takes precedence, and $ref is ignored.
Resolution of ref at property level is not needed in that scenario.